Update pending txs when last block is updated
